# docs/generated/openapi.yaml
openapi: 3.0.1
info:
  title: Blog API
  description: |-
    Auto-generated OpenAPI spec via springdoc-openapi.

    - Web APIs: /api
    - Admin APIs: /api/admin

    Admin APIs require: Authorization: Bearer <token>
  version: v1
servers:
- url: http://localhost:8080
  description: Local dev
security:
- bearer-jwt: []
...
paths:
...
  /api/admin/comments/{id}/reject:
    put:
      tags:
      - Admin - Comments
      summary: Reject comment
      operationId: reject
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/admin/comments/{id}/approve:
    put:
      tags:
      - Admin - Comments
      summary: Approve comment
      operationId: approve
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
  /api/admin/categories/{id}:
    put:
      tags:
      - Admin - Categories
      summary: Update category
      operationId: update_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameOnlyRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
    delete:
      tags:
      - Admin - Categories
      summary: Delete category
      operationId: delete_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
...
  /api/posts/{id}/comments:
    get:
      tags:
      - Web - Comments
      summary: List approved comments
      description: Lists comments for a post (APPROVED only).
      operationId: list_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListCommentVO'
    post:
      tags:
      - Web - Comments
      summary: Create comment
      description: Creates a comment for a post. New comments default to PENDING for
        admin review.
      operationId: create
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLong'
...
  /api/admin/categories:
    get:
      tags:
      - Admin - Categories
      summary: List categories
      operationId: list_5
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListCategoryVO'
    post:
      tags:
      - Admin - Categories
      summary: Create category
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameOnlyRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseLong'
...
  /api/posts:
    get:
      tags:
      - Web - Posts
      summary: List published posts
      description: Supports pagination and optional filters (keyword/categoryId/tagId).
      operationId: list_1
      parameters:
      - name: q
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/PostQuery'
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageResultPostListItemVO'
...
  /api/admin/posts/meta/categories:
    get:
      tags:
      - Admin - Posts
      summary: List categories for editor
      operationId: categories
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListCategoryVO'
...
  /api/admin/comments:
    get:
      tags:
      - Admin - Comments
      summary: List comments
      description: Admin comment list with pagination and optional filters (status/postId).
      operationId: page_2
      parameters:
      - name: q
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/AdminCommentQuery'
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePageResultAdminCommentVO'
...
  /api/admin/comments/{id}:
    delete:
      tags:
      - Admin - Comments
      summary: Delete comment
      operationId: delete_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: object
          description: Response data payload
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    NameOnlyRequest:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 50
          minLength: 0
          type: string
          description: Name
          example: Java
...
    CommentCreateRequest:
      required:
      - content
      - nickname
      type: object
      properties:
        nickname:
          maxLength: 30
          minLength: 0
          type: string
          description: Nickname displayed on comment
          example: Tom
        email:
          maxLength: 100
          minLength: 0
          type: string
          description: "Email (optional). Used for contact only, not displayed publicly"
          example: tom@example.com
        content:
          maxLength: 500
          minLength: 0
          type: string
          description: Comment content
          example: Nice post!
    ApiResponseLong:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: integer
          description: Response data payload
          format: int64
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
...
    TagVO:
      type: object
      properties:
        id:
          type: integer
          description: Tag id
          format: int64
          example: 1
        name:
          type: string
          description: Tag name
          example: Spring
      description: Tag
...
    PostQuery:
      type: object
      properties:
        pageNum:
          minimum: 1
          type: integer
          description: Page number (start from 1)
          format: int64
          example: 1
        pageSize:
          maximum: 100
          minimum: 1
          type: integer
          description: Page size (1~100)
          format: int64
          example: 10
        keyword:
          maxLength: 50
          minLength: 0
          type: string
          description: Optional keyword for title/summary search
          example: spring
        categoryId:
          type: integer
          description: Filter by categoryId
          format: int64
          example: 1
        tagId:
          type: integer
          description: Filter by tagId
          format: int64
          example: 2
    ApiResponsePageResultPostListItemVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          $ref: '#/components/schemas/PageResultPostListItemVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    CategoryVO:
      type: object
      properties:
        id:
          type: integer
          description: Category id
          format: int64
          example: 1
        name:
          type: string
          description: Category name
          example: Java
      description: Category
    PageResultPostListItemVO:
      type: object
      properties:
        list:
          type: array
          description: Current page list
          items:
            $ref: '#/components/schemas/PostListItemVO'
        total:
          type: integer
          description: Total items count
          format: int64
          example: 100
        pageNum:
          type: integer
          description: Current page number
          format: int64
          example: 1
        pageSize:
          type: integer
          description: Page size
          format: int64
          example: 10
      description: Pagination result wrapper
    PostListItemVO:
      type: object
      properties:
        id:
          type: integer
          description: Post id
          format: int64
          example: 1
        title:
          type: string
          description: Post title
          example: Hello World
        summary:
          type: string
          description: Post summary
          example: Short introduction
        coverUrl:
          type: string
          description: Cover image URL
          example: /uploads/202602/cover.jpg
        category:
          $ref: '#/components/schemas/CategoryVO'
        tags:
          type: array
          description: Tags
          items:
            $ref: '#/components/schemas/TagVO'
        publishedAt:
          type: string
          description: Published time (string)
          example: 2026-02-01 12:00:00
        viewCount:
          type: integer
          description: View count
          format: int64
          example: 123
      description: Post list item
...
    ApiResponseListCommentVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: array
          description: Response data payload
          items:
            $ref: '#/components/schemas/CommentVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    CommentVO:
      type: object
      properties:
        id:
          type: integer
          description: Comment id
          format: int64
          example: 1
        nickname:
          type: string
          description: Nickname
          example: Tom
        content:
          type: string
          description: Content
          example: Nice post!
        createdAt:
          type: string
          description: Created time
          example: 2026-02-01 12:00:00
      description: Comment
...
    ApiResponseListCategoryVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          type: array
          description: Response data payload
          items:
            $ref: '#/components/schemas/CategoryVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
...
    AdminCommentQuery:
      type: object
      properties:
        pageNum:
          minimum: 1
          type: integer
          description: Page number (start from 1)
          format: int64
          example: 1
        pageSize:
          maximum: 100
          minimum: 1
          type: integer
          description: Page size (1~100)
          format: int64
          example: 10
        status:
          maxLength: 20
          minLength: 0
          type: string
          description: "Comment status (optional): PENDING/APPROVED/REJECTED"
          example: PENDING
          enum:
          - PENDING
          - APPROVED
          - REJECTED
        postId:
          type: integer
          description: Filter by postId
          format: int64
          example: 1
    AdminCommentVO:
      type: object
      properties:
        id:
          type: integer
          description: Comment id
          format: int64
          example: 1
        postId:
          type: integer
          description: Post id
          format: int64
          example: 1
        postTitle:
          type: string
          description: Post title
          example: Hello World
        nickname:
          type: string
          description: Nickname
          example: Tom
        email:
          type: string
          description: Email
          example: tom@example.com
        content:
          type: string
          description: Content
          example: Nice post!
        status:
          type: string
          description: Status
          example: PENDING
          enum:
          - PENDING
          - APPROVED
          - REJECTED
        createdAt:
          type: string
          description: Created time
          example: 2026-02-01 12:00:00
      description: Admin comment item
    ApiResponsePageResultAdminCommentVO:
      type: object
      properties:
        code:
          type: integer
          description: "Business status code. 0 means ok, non-zero means error."
          format: int32
          example: 0
        message:
          type: string
          description: Message for humans
          example: ok
        data:
          $ref: '#/components/schemas/PageResultAdminCommentVO'
        timestamp:
          type: integer
          description: Server timestamp in milliseconds
          format: int64
          example: 1700000000000
      description: Standard API response wrapper
    PageResultAdminCommentVO:
      type: object
      properties:
        list:
          type: array
          description: Current page list
          items:
            $ref: '#/components/schemas/AdminCommentVO'
        total:
          type: integer
          description: Total items count
          format: int64
          example: 100
        pageNum:
          type: integer
          description: Current page number
          format: int64
          example: 1
        pageSize:
          type: integer
          description: Page size
          format: int64
          example: 10
      description: Pagination result wrapper
...
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT